+++ docs/changes.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Version 0.8.0
+-------------
+
+To be released.
+
+
 Version 0.7.0
 -------------